requirements.txt
pyarrow
google-cloud-storage
pandas
pytest
functions_framework
gcsfs